moves new suggestions system to images module
