+++ lib/modules/apostrophe-images/index.js
@@ -162,6 +162,7 @@ module.exports = {
     self.pushAsset('script', 'focal-point-editor', { when: 'user' });
     self.pushAsset('stylesheet', 'user', { when: 'user' });
     require('./lib/api.js')(self, options);
+    require('./lib/routes.js')(self, options);
     self.enableHelpers();
   }
 };
