+++ lib/modules/apostrophe-images/lib/api.js
@@ -151,4 +151,40 @@ module.exports = function(self, options) {
     return projection;
   };
 
+  self.getSuggestions = async function(req, term) {
+    const search = self.apos.launder.string(req.body.term).toLowerCase();
+    const searchReg = new RegExp(`^${search}`, 'i');
+
+    const images = await self.find(req, {
+      $or: [
+        {
+          title: searchReg
+        },
+        {
+          tags: {
+            $in: [searchReg]
+          }
+        }
+      ]
+    }).limit(10).toArray();
+    // .autocomplete(search)
+
+    const suggestions = images.reduce((acc, img) => {
+      if (acc.length >= 10) {
+        return acc;
+      }
+
+      const title = img.title.toLowerCase();
+      const titleMerge = (title.includes(search) && !acc.includes(title)) ? [ img.title ] : [];
+      const tagsMerge = img.tags.filter((tag) => !acc.includes(tag) && tag.startsWith(search));
+
+      return [
+        ...acc,
+        ...titleMerge,
+        ...tagsMerge
+      ];
+    }, []);
+
+    return suggestions;
+  };
 };
