+++ lib/modules/apostrophe-pieces/lib/api.js
@@ -124,37 +124,42 @@ module.exports = function(self, options) {
     return next();
   };
 
-  self.getSuggestions = async function(req, term) {
-    const search = self.apos.launder.string(req.body.term).toLowerCase();
+  // self.getSuggestions = async function(req, term) {
+  //   const search = self.apos.launder.string(req.body.term).toLowerCase();
+  //   const searchReg = new RegExp(`^${search}`, 'i');
 
-    const images = await self.find(req, {
-      // $or: [
-      // {
-      //   title: {$regex: `^${term}`}
-      // }
-      // {
-      //   tags: {
-      //     $in: [termReg]
-      //   }
-      // }
-      // ]
-      // highSearchText: self.apos.utils.searchify(term)
-    }).limit(10).autocomplete(search).toArray();
+  //   const images = await self.find(req, {
+  //     $or: [
+  //       {
+  //         title: searchReg
+  //       },
+  //       {
+  //         tags: {
+  //           $in: [searchReg]
+  //         }
+  //       }
+  //     ]
+  //   }).limit(10).toArray();
+  //   // .autocomplete(search)
 
-    const suggestions = images.reduce((acc, img) => {
-      const title = img.title.toLowerCase();
-      const titleMerge = (title.includes(search) && !acc.includes(title)) ? [ img.title ] : [];
-      const tagsMerge = img.tags.filter((tag) => !acc.includes(tag) && tag.startsWith(search));
+  //   const suggestions = images.reduce((acc, img) => {
+  //     if (acc.length >= 10) {
+  //       return acc;
+  //     }
 
-      return [
-        ...acc,
-        ...titleMerge,
-        ...tagsMerge
-      ];
-    }, []);
+  //     const title = img.title.toLowerCase();
+  //     const titleMerge = (title.includes(search) && !acc.includes(title)) ? [ img.title ] : [];
+  //     const tagsMerge = img.tags.filter((tag) => !acc.includes(tag) && tag.startsWith(search));
 
-    return suggestions;
-  };
+  //     return [
+  //       ...acc,
+  //       ...titleMerge,
+  //       ...tagsMerge
+  //     ];
+  //   }, []);
+
+  //   return suggestions;
+  // };
 
   // options.filters can contain cursor filters. `options.chooser`, `options.format` and
   // `options.manageView` are also implemented. For bc, if `options.filters` does not exist,
