+++ lib/modules/apostrophe-pieces/lib/routes.js
@@ -7,7 +7,7 @@ module.exports = function(self, options) {
 
     self.route('post', 'insert', self.requireEditor, self.routes.insert);
     self.route('post', 'retrieve', self.requirePieceEditorView, self.routes.retrieve);
-    self.route('post', 'suggestions', self.apos.docs.requireEditorOfSomething, self.routes.suggestions);
+    // self.route('post', 'suggestions', self.apos.docs.requireEditorOfSomething, self.routes.suggestions);
     self.route('post', 'list', self.apos.docs.requireEditorOfSomething, self.routes.list);
     self.route('post', 'update', self.requirePiece, self.routes.update);
     self.route('post', 'publish', self.requireEditor, self.routes.publish);
@@ -35,15 +35,15 @@ module.exports = function(self, options) {
     return self.retrieveResponse(req, res, null, req.piece);
   };
 
-  self.routes.suggestions = async function(req, res, next) {
-    if (!req.body.term) {
-      return res.status(200).send();
-    }
+  // self.routes.suggestions = async function(req, res, next) {
+  //   if (!req.body.term) {
+  //     return res.status(200).send();
+  //   }
 
-    const suggestions = await self.getSuggestions(req, req.body.term);
+  //   const suggestions = await self.getSuggestions(req, req.body.term);
 
-    res.status(200).send(suggestions);
-  };
+  //   res.status(200).send(suggestions);
+  // };
 
   self.routes.list = function(req, res) {
     var results;
