+++ lib/modules/apostrophe-pieces/public/js/manager-modal.js
@@ -40,8 +40,8 @@ apos.define('apostrophe-pieces-manager-modal', {
       self.enableChooseViews();
       self.enableFilters();
       self.enableSorts();
-      self.enableSuggestions();
-      // self.enableSearch();
+      // self.enableSuggestions();
+      self.enableSearch();
       self.enableCheckboxEvents();
       self.enableBatchOperations();
       self.enableInsertViaUpload();
@@ -620,132 +620,136 @@ apos.define('apostrophe-pieces-manager-modal', {
       });
     };
 
-    self.enableSuggestions = function() {
-      var isListening = false;
-      var inputSelector = '[name="search-' + self.options.name + '"]';
+    // self.enableSuggestions = function() {
+    //   var isListening = false;
+    //   var inputSelector = '[name="search-' + self.options.name + '"]';
 
-      self.$filters.on('keyup', inputSelector, function(e) {
-        var $input = $(e.target);
-        var $search = self.$filters.find('.apos-modal-filters-search');
-        var $suggestionsList = $search.find('.ui-autocomplete');
-        var $activeItem = $suggestionsList.find('li.ui-menu-item.active');
-        var listener = getListener($search);
+    //   self.$filters.on('keyup', inputSelector, function(e) {
+    //     var $input = $(e.target);
+    //     var $search = self.$filters.find('.apos-modal-filters-search');
+    //     var $suggestionsList = $search.find('.ui-autocomplete');
+    //     var $activeItem = $suggestionsList.find('li.ui-menu-item.active');
+    //     var listener = getListener($search);
 
-        if (e.keyCode === 13) {
-          return search($input, $activeItem);
-        }
+    //     if (e.keyCode === 13) {
+    //       return search($input, $activeItem);
+    //     }
 
-        if ([38, 40].includes(e.keyCode)) {
-          return selectItems($suggestionsList, $activeItem, e.keyCode);
-        }
+    //     if ([38, 40].includes(e.keyCode)) {
+    //       return selectItems($suggestionsList, $activeItem, e.keyCode);
+    //     }
 
-        var options = {
-          term: $input.val()
-        };
+    //     var options = {
+    //       term: $input.val()
+    //     };
 
-        self.api('suggestions', options, function(results) {
-          if (results && results.length) {
-            var suggestions = '<ul class="ui-autocomplete">';
-            results.forEach(function (result, i) {
-              suggestions = suggestions.concat('<li class="ui-menu-item">' + result + '</li>');
-              if (i === results.length - 1) {
-                suggestions = suggestions.concat('</ul>');
-              }
-            });
+    //     self.api('suggestions', options, function(results) {
+    //       if (results && results.length) {
+    //         var suggestions = '<ul class="ui-autocomplete">';
+    //         results.forEach(function (result, i) {
+    //           suggestions = suggestions.concat('<li class="ui-menu-item">' + result + '</li>');
+    //           if (i === results.length - 1) {
+    //             suggestions = suggestions.concat('</ul>');
+    //           }
+    //         });
 
-            $search.find('ul').remove();
-            $search.append(suggestions);
+    //         $search.find('ul').remove();
+    //         $search.append(suggestions);
 
-            $search.find('ul > .ui-menu-item').each(function (i, elem) {
-              var $el = $(elem);
+    //         $search.find('ul > .ui-menu-item').each(function (i, elem) {
+    //           var $el = $(elem);
 
-              $(elem).on('click', function() {
-                search($el);
-              });
-            });
+    //           $(elem).on('click', function() {
+    //             search($input, $el);
+    //           });
+    //         });
 
-            if (!isListening) {
-              isListening = true;
-              window.addEventListener('click', listener);
-            }
+    //         if (!isListening) {
+    //           isListening = true;
+    //           window.addEventListener('click', listener);
+    //         }
 
-          } else {
-            if (isListening) {
-              listener();
-            }
-          }
-        });
-      });
-
-      function search ($el, $activeItem) {
-        var val;
-        if ($activeItem && $activeItem.length) {
-          val = $activeItem.text();
-        } else {
-          val = $el.text();
-        }
-
-        self.search = val;
-        self.currentFilters.page = 1;
-        self.refresh(function() {
-          // refocus input element after search
-          $el.focus();
-          // put the search query before the cursor
-          $el.val(val);
-        });
-      };
-
-      function getListener ($search) {
-        return function listener () {
-          $search.find('ul').remove();
-          isListening = false;
-          window.removeEventListener('click', listener);
-        };
-      };
-
-      function selectItems ($suggestionsList, $activeItem, keyCode) {
-        if (!$suggestionsList.length) {
-          return;
-        }
-
-        var isDown = keyCode === 40;
-        var firstOrLastSelector = ':last-child';
-        if (isDown) {
-          firstOrLastSelector = ':first-child';
-        }
-
-        if ($activeItem.length) {
-          var $nextItem;
-          if (isDown) {
-            $nextItem = $activeItem.next();
-          } else {
-            $nextItem = $activeItem.prev();
-          }
-
-          if ($nextItem.length) {
-            $nextItem.focus().addClass('active');
-          }
-          $activeItem.removeClass('active');
-
-        } else {
-          $suggestionsList.find('li' + firstOrLastSelector).focus().addClass('active');
-        }
-      }
-    };
-
-    // self.enableSearch = function() {
-    //   self.$filters.on('change', '[name="search-' + self.options.name + '"]', function() {
-    //     self.search = $(this).val();
-    //     self.currentFilters.page = 1;
-    //     self.refresh(function() {
-    //       var $input = self.$filters.find('[name="search-' + self.options.name + '"]');
-    //       // refocus input element after search
-    //       $input.focus();
-    //       // put the search query before the cursor
-    //       $input.val($input.val());
+    //       } else {
+    //         if (isListening) {
+    //           listener();
+    //         }
+    //       }
     //     });
     //   });
+
+    //   function search ($input, $activeItem) {
+    //     var val;
+    //     if ($activeItem && $activeItem.length) {
+    //       val = $activeItem.text();
+    //     } else {
+    //       val = $input.text();
+    //     }
+
+    //     self.search = val;
+    //     self.currentFilters.page = 1;
+    //     $input.focus();
+
+    //     self.refresh(function() {
+    //       // refocus input element after search
+    //       console.log('$input ===> ', $input);
+    //       console.log('val ===> ', val);
+    //       $input.focus();
+    //       // put the search query before the cursor
+    //       $input.val(val);
+    //     });
+    //   };
+
+    //   function getListener ($search) {
+    //     return function listener () {
+    //       $search.find('ul').remove();
+    //       isListening = false;
+    //       window.removeEventListener('click', listener);
+    //     };
+    //   };
+
+    //   function selectItems ($suggestionsList, $activeItem, keyCode) {
+    //     if (!$suggestionsList.length) {
+    //       return;
+    //     }
+
+    //     var isDown = keyCode === 40;
+    //     var firstOrLastSelector = ':last-child';
+    //     if (isDown) {
+    //       firstOrLastSelector = ':first-child';
+    //     }
+
+    //     if ($activeItem.length) {
+    //       var $nextItem;
+    //       if (isDown) {
+    //         $nextItem = $activeItem.next();
+    //       } else {
+    //         $nextItem = $activeItem.prev();
+    //       }
+
+    //       if ($nextItem.length) {
+    //         $nextItem.focus().addClass('active');
+    //       }
+    //       $activeItem.removeClass('active');
+
+    //     } else {
+    //       $suggestionsList.find('li' + firstOrLastSelector).focus().addClass('active');
+    //     }
+    //   }
     // };
+
+    self.enableSearch = function() {
+      self.$filters.on('change', '[name="search-' + self.options.name + '"]', function() {
+        self.search = $(this).val();
+        self.currentFilters.page = 1;
+        self.refresh(function() {
+          var $input = self.$filters.find('[name="search-' + self.options.name + '"]');
+          // refocus input element after search
+          $input.focus();
+          // put the search query before the cursor
+          $input.val($input.val());
+        });
+      });
+    };
 
     self.choices = [];
